Moved validateFlow and tests to own class, refactored readHeader method in validate flow and moved to DicomParsing
